+++ src/serve.ts
@@ -1,42 +1,56 @@
 import fs from 'fs'
-import { type ViteDevServer } from 'vite'
+import { normalizePath, type ViteDevServer } from 'vite'
 import {
   type Configuration,
   resolveConfig,
 } from './config'
 import { build } from './build'
+import { ensuredir, STATIC_JS_EXTENSIONS } from './utils'
 
 export async function bootstrap(config: Configuration, server: ViteDevServer) {
   process.env.VITE_DEV_SERVER_URL = resolveEnv(server)!.url
 
   const resolved = await resolveConfig(config, 'serve', server)
   const {
+    _fn,
     watcher,
-    src2dist,
-    config: rawConfig,
-    startup,
-    isPreload,
+    plugins,
+    extensions,
   } = resolved
-  const startup_fn = debounce(function startup_fn(filename: string) {
-    if (isPreload(filename)) {
-      server.ws.send({ type: 'full-reload' })
-    } else {
-      startup()
-    }
-  })
   // There can't be any await statement here, it will cause `watcher.on` to miss the first trigger.
-  watcher!.on('all', async (event, filepath) => {
-    // 🚨 filepath = normalizePath(filepath)
-    rawConfig.onwatch?.(event, filepath)
+  watcher!.on('all', async (event, _filepath) => {
+    const filepath = normalizePath(_filepath)
+    const distpath = _fn.include2dist(filepath)
 
-    const distpath = src2dist(filepath)
+    // call onwatch hooks
+    for (const plugin of plugins) {
+      plugin.onwatch?.(event, filepath)
+    }
+
     switch (event) {
       case 'add':
-      case 'change':
-        await build(resolved, filepath)
+      case 'change': {
+        const filename = filepath
+        const js = extensions.some(ext => filename.endsWith(ext))
+        const staticjs = STATIC_JS_EXTENSIONS.some(ext => filename.endsWith(ext))
+        const distname = _fn.include2dist(filename, js)
+        if (js) {
+          await build(resolved, filename)
+        } else if (staticjs) {
+          // static files
+          fs.copyFileSync(filename, ensuredir(distname))
+        }
+
+        if (js || staticjs) {
+          for (const plugin of plugins) {
+            // call ondone hooks
+            plugin.ondone?.({ filename, distname })
+          }
+        }
         break
+      }
       case 'addDir':
-        !fs.existsSync(distpath) && fs.mkdirSync(distpath, { recursive: true })
+        // !fs.existsSync(distpath) && fs.mkdirSync(distpath, { recursive: true })
         break
       case 'unlink':
         fs.existsSync(distpath) && fs.unlinkSync(distpath)
@@ -45,23 +59,7 @@ export async function bootstrap(config: Configuration, server: ViteDevServer) {
         fs.existsSync(distpath) && fs.rmSync(distpath, { recursive: true, force: true })
         break
     }
-
-    if (rawConfig.startup !== false) {
-      startup_fn(filepath)
-    }
   })
-
-  // first start
-  // watcher!.emit('add', 'add', config.include2files[0])
-}
-
-function debounce<Fn extends (...args: any[]) => void>(fn: Fn, delay = 299) {
-  let t: NodeJS.Timeout
-  return ((...args) => {
-    // !t && fn(...args) // first call
-    clearTimeout(t)
-    t = setTimeout(() => fn(...args), delay)
-  }) as Fn
 }
 
 /**